Updates infrastructure to use official Vue multi-page pattern.
