+++ prereq_map/static/prereq_map/js/components/course-detail.vue
@@ -52,7 +52,9 @@
 
 <script>
 import Graph from "./course-graph.vue";
-import { dataBus } from "../course";
+
+import { dataBus } from "../pages/course/";
+
 export default {
     components: {
         'course-graph': Graph
