+++ prereq_map/templates/base.html
@@ -72,6 +72,7 @@
 
     <main class="prereq-content">
         {% block content %}{% endblock %}
+
     </main>
 
     <footer class="prereq-footer" role="contentinfo">
