+++ prereq_map/templates/course.html
@@ -7,28 +7,8 @@
 
 {% block content %}
 
-<div id="vue_course" class="container py-1 mt-2" role="main">
-
-    <h1 class="pt-4 pb-2">Course Search</h1>
-    <p class="instruction-text pb-4">View course prerequisites and related curricula</p>
-
-    <course-search-input></course-search-input>
-
-    <div v-cloak v-if="course_param !== undefined">
-
-        <div v-if="course_valid" class="row mt-5 mb-5">
-            <course-detail v-bind:course-param="course_param"></course-detail>
-        </div>
-
-        <div v-cloak v-if="course_valid === false" class="row mt-5 mb-5">
-            The course you entered does not exist or has no data
-        </div>
-
-    </div>
-
-</div>
-
-<!-- load bundled vue files using webpack loader -->
-{% render_bundle 'vue_course' %}
+<!-- vue page components will be injected here -->
+<div id="course"></div>
+{% render_bundle 'course' %}
 
 {% endblock %}
