+++ webpack.config.js
@@ -11,9 +11,9 @@ module.exports = {
     context: __dirname,
 
     entry: {
-        base: './prereq_map/static/prereq_map/js/base',
-        vue_curriculum: './prereq_map/static/prereq_map/js/curriculum',
-        vue_course: './prereq_map/static/prereq_map/js/course'
+        base: './prereq_map/static/prereq_map/js/base.js',
+        curriculum: './prereq_map/static/prereq_map/js/pages/curriculum/',
+        course: './prereq_map/static/prereq_map/js/pages/course/'
     },
 
     optimization: {
